+++ lib/components/Button/styles.ts
@@ -8,72 +8,76 @@ import borderRadius from '@/tokens/future/borderRadius.json'
 import fonts from '@/tokens/future/fonts.json'
 import shadows from '@/tokens/future/shadows.json'
 
+const sm = fonts['button-small']
+const md = fonts['button-medium']
+const lg = fonts['button-large']
+
 export default createUseStyles({
   btn: {
-    display: 'inline-block', //*
-    boxSizing: 'border-box', //*
-    position: 'relative', //*
-    maxWidth: '100%', //*
-    marginBottom: 0, //*
-    padding: [spacing['size-3'], spacing['size-4']], //*
-    borderRadius: borderRadius['br-xs'], //*
+    display: 'inline-block',
+    boxSizing: 'border-box',
+    position: 'relative',
+    maxWidth: '100%',
+    marginBottom: 0,
+    padding: [spacing['size-3'], spacing['size-4']],
+    borderRadius: borderRadius['br-xs'],
     border: 0,
-    font: fonts['button-small'], //*
-    textAlign: 'center', //*
-    textDecoration: 'none', //*
-    whiteSpace: 'nowrap', //*
+    font: `${sm.fontWeight} ${sm.fontSize}/${sm.lineHeight} ${sm.fontFamily}`,
+    textAlign: 'center',
+    textDecoration: 'none',
+    whiteSpace: 'nowrap',
     transition: '0.3s all',
-    cursor: 'pointer', //*
-    userSelect: 'none', //*
-    touchAction: 'manipulation', //*
+    cursor: 'pointer',
+    userSelect: 'none',
+    touchAction: 'manipulation',
     '&:focus': {
-      outline: 'none' //*
+      outline: 'none'
     },
     '&:hover': {
-      textDecoration: 'none' //*
+      textDecoration: 'none'
     }
   },
   cont: {
-    display: 'flex', //*
-    width: '100%', //*
-    height: '100%', //*
-    alignItems: 'center', //*
-    justifyContent: 'center', //*
-    pointerEvents: 'none' //*
+    display: 'flex',
+    width: '100%',
+    height: '100%',
+    alignItems: 'center',
+    justifyContent: 'center',
+    pointerEvents: 'none'
   },
   loadCont: {
-    display: 'flex', //*
-    width: '100%', //*
-    height: '100%', //*
-    alignItems: 'center', //*
-    justifyContent: 'center', //*
-    position: 'absolute', //*
-    top: 0, //*
-    left: 0, //*
-    pointerEvents: 'none' //*
+    display: 'flex',
+    width: '100%',
+    height: '100%',
+    alignItems: 'center',
+    justifyContent: 'center',
+    position: 'absolute',
+    top: 0,
+    left: 0,
+    pointerEvents: 'none'
   },
   // Themes
   primary: {
-    background: colors.button.primary.bg.default, //*
-    color: colors.text.white.primary, //*
-    outline: `2px solid ${colors.button.primary.border.default}`, //*
-    outlineOffset: '-2px', //*
+    background: colors.button.primary.bg.default,
+    color: colors.text.white.primary,
+    outline: `2px solid ${colors.button.primary.border.default}`,
+    outlineOffset: '-2px',
     '&:hover': {
-      background: colors.button.primary.bg.hover //*
+      background: colors.button.primary.bg.hover
     },
     '&:active': {
-      background: colors.button.primary.bg.active //*
+      background: colors.button.primary.bg.active
     },
     '&:focus': {
-      outline: `2px solid ${colors.button.primary.border.default}` //*
+      outline: `2px solid ${colors.button.primary.border.default}`
     },
     '&:not(:active):focus-visible': {
-      boxShadow: shadows['focus-pink'] //*
+      boxShadow: shadows['focus-pink']
     },
     '&$disabled': {
-      background: colors.button.primary.bg.disabled, //*
-      color: colors.text.white.secondary, //*
-      cursor: 'not-allowed' //*
+      background: colors.button.primary.bg.disabled,
+      color: colors.text.white.secondary,
+      cursor: 'not-allowed'
     },
     '& $icon': {
       fill: colors.icon.inverse.default
@@ -83,20 +87,20 @@ export default createUseStyles({
     }
   },
   secondary: {
-    background: colors.button.secondary.bg.default, //*
-    color: colors.text.indigo.primary, //*
+    background: colors.button.secondary.bg.default,
+    color: colors.text.indigo.primary,
     '&:hover': {
-      background: colors.button.secondary.bg.hover //*
+      background: colors.button.secondary.bg.hover
     },
     '&:active': {
-      background: colors.button.secondary.bg.active //*
+      background: colors.button.secondary.bg.active
     },
     '&:not(:active):focus-visible': {
-      boxShadow: shadows['focus-indigo'] //*
+      boxShadow: shadows['focus-indigo']
     },
     '&$disabled': {
-      background: colors.button.secondary.bg.disabled, //*
-      color: colors.text.indigo.secondary //*
+      background: colors.button.secondary.bg.disabled,
+      color: colors.text.indigo.secondary
     },
     '& $icon': {
       fill: colors.icon.brand.default
@@ -106,26 +110,26 @@ export default createUseStyles({
     }
   },
   tertiary: {
-    background: 'transparent', //*
-    color: colors.text.indigo.primary, //*
-    outline: `2px solid ${colors.button.tertiary.border.default}`, //*
-    outlineOffset: '-2px', //*
+    background: 'transparent',
+    color: colors.text.indigo.primary,
+    outline: `2px solid ${colors.button.tertiary.border.default}`,
+    outlineOffset: '-2px',
     '&:hover': {
-      background: colors.button.tertiary.bg.hover //*
+      background: colors.button.tertiary.bg.hover
     },
     '&:active': {
-      background: colors.button.tertiary.bg.active //*
+      background: colors.button.tertiary.bg.active
     },
     '&:focus': {
-      outline: `2px solid ${colors.button.tertiary.border.default}` //*
+      outline: `2px solid ${colors.button.tertiary.border.default}`
     },
     '&:not(:active):focus-visible': {
-      boxShadow: shadows['focus-indigo'] //*
+      boxShadow: shadows['focus-indigo']
     },
     '&$disabled': {
-      background: 'transparent', //*
-      color: colors.text.indigo.secondary, //*
-      outline: `2px solid ${colors.button.tertiary.border.disabled}` //*
+      background: 'transparent',
+      color: colors.text.indigo.secondary,
+      outline: `2px solid ${colors.button.tertiary.border.disabled}`
     },
     '& $icon': {
       fill: colors.icon.brand.default
@@ -238,23 +242,23 @@ export default createUseStyles({
   },
   // Sizes
   md: {
-    font: fonts['button-medium'], //*
-    padding: [spacing['size-3'], spacing['size-5']], //*
+    font: `${md.fontWeight} ${md.fontSize}/${md.lineHeight} ${md.fontFamily}`,
+    padding: [spacing['size-3'], spacing['size-5']],
     '&$iconOnly': {
       padding: [spacing['size-2'], spacing['size-2']]
     }
   },
   lg: {
-    font: fonts['button-large'], //*
-    padding: [spacing['size-4'], spacing['size-6']], //*
+    font: `${lg.fontWeight} ${lg.fontSize}/${lg.lineHeight} ${lg.fontFamily}`,
+    padding: [spacing['size-4'], spacing['size-6']],
     '&$iconOnly': {
       padding: [spacing['size-3'], spacing['size-3']]
     }
   },
   // Block
   block: {
-    display: 'block', //*
-    width: '100%' //*
+    display: 'block',
+    width: '100%'
   },
   // Icon
   icon: {},