+++ src/components/MDXComponents.ts
@@ -81,7 +81,8 @@ const mdxComponents = {
       tag: 'li',
       children: props.children
     }),
-  a: Anchor
+  a: Anchor,
+  blockquote: Alert
 }
 
 export default mdxComponents